refactor block storage config
