+++ config/config.go
@@ -3,6 +3,7 @@ package config
 import (
 	"github.com/orbs-network/orbs-spec/types/go/primitives"
 	"github.com/orbs-network/orbs-spec/types/go/protocol/consensus"
+	"time"
 )
 
 type NodeConfig interface {
@@ -10,4 +11,5 @@ type NodeConfig interface {
 	NetworkSize(asOfBlock uint64) uint32
 	ConstantConsensusLeader() primitives.Ed25519PublicKey
 	ActiveConsensusAlgo() consensus.ConsensusAlgoType
+	BlockSyncCommitTimeout() time.Duration
 }
