+++ test/e2e/e2e_test.go
@@ -57,12 +57,13 @@ var _ = Describe("The Orbs Network", func() {
 				":8080",
 				nodePublicKey,
 				1,
+				70,
 				constantConsensusLeaderPublicKey,
 				consensus.CONSENSUS_ALGO_TYPE_LEAN_HELIX,
 				gossipTransport,
 			)
-      
-      // To let node start up properly, otherwise in Docker we get connection refused
+
+			// To let node start up properly, otherwise in Docker we get connection refused
 			time.Sleep(100 * time.Millisecond)
 		}
